added sample excuse feature as well as reorganized data
