+++ src/data/excusesData.py
@@ -1,0 +1,26 @@
+sample_excuses = [
+    "I thought I saw a squirrel outside and got distracted.",
+    "I got caught up in an epic daydream about the meaning of life.",
+    "I was busy reorganizing my sock drawer—it's a work in progress!",
+    "I suddenly remembered I left the oven on.",
+    "I was busy perfecting my procrastination skills.",
+    "I got stuck in an endless scroll of funny cat videos.",
+    "I spent too much time coming up with excuses instead of doing the work.",
+    "I had a brilliant idea for a new hobby and lost track of time.",
+    "I was too busy debating with my pet goldfish.",
+    "I got lost in a maze of memes.",
+    "I had to reorganize my thoughts; it took longer than expected.",
+    "I accidentally hit the snooze button on my to-do list.",
+    "I was busy rethinking the entire plan.",
+    "My brain decided to go on a coffee break.",
+    "I got caught up in a spontaneous dance party."
+]
+
+custom_excuse_templates = [
+    "I can't do {task} right now because I need to finish my snack.",
+    "I'm too busy to {task} at the moment; my cat needs attention.",
+    "I would love to {task}, but I have to binge-watch my favorite show first.",
+    "I can't {task} right now; I'm waiting for the pizza delivery.",
+    "I need to take a break from {task} to recharge my brain.",
+    "I can't focus on {task} because I'm too busy daydreaming.",
+]
