+++ src/get_excuse.py
@@ -1,0 +1,4 @@
+import random
+from data.excusesData import sample_excuses
+def get_excuse():
+    return random.choice(sample_excuses)
